+++ apps/api/config/initializers/validate_supplier_credentials.rb
@@ -5,6 +5,7 @@ if enforce_on_envs.include?(Hanami.env)
     atleisure:  %w(username password test_mode),
     jtb:        %w(id user password company url),
     kigo:       %w(subscription_key),
-    kigolegacy: %w(username password)
+    kigolegacy: %w(username password),
+    waytostay:  %w(client_id client_secret url token_url)
   })
 end